backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-rest-auth>=7.2.0",
    "django>=6.0.3",
    "django-allauth>=65.15.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.17.1",
    "google-genai>=1.70.0",
    "instructor>=1.15.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "coverage>=7.13.5",
    "django-stubs>=6.0.2",
    "djangorestframework-stubs>=3.16.9",
    "mypy>=1.20.0",
    "ruff>=0.15.9",
    "types-requests>=2.33.0.20260402",
]
